Added custom message factory support
Users can use this to change the DefaultGelfMessageFactory behaviour

Example:
    <appender name="graylogMessages" class="com.github.pukkaone.gelf.logback.GelfAppender">
        ...
        <marshaller class="my.custom.package..MyGelfMessageFactory">
            <mdcTimestampProperty>ReceivedDateTime</mdcTimestampProperty>
        </marshaller>
        ...
    </appender>
